proj/requirements.txt
django-filter
django
djangorestframework
celery
flower
celery[redis]
markdown
mongoengine
psycopg2